CLDSRV-805: Regroup object tagging into 1 test
Using 1 bucket on test suite to reduce gcp quota pressure
